Move remaining Store.update_ methods to updater
